+++ R/fit-model.R
@@ -4,7 +4,7 @@
 #' @param proj.end end year for projection.
 #' @param popupdate logical should target population be updated to match
 #'   age-specific population size from DP file and %Urban from EPP XML.
-prepare_spec_fit <- function(pjnz, proj.end=2016.5, popadjust = TRUE, popupdate=TRUE){
+prepare_spec_fit <- function(pjnz, proj.end=2016.5, popadjust = TRUE, popupdate=TRUE, use_ep5=FALSE){
 
   ## epp
   eppd <- read_epp_data(pjnz)
@@ -14,8 +14,8 @@ prepare_spec_fit <- function(pjnz, proj.end=2016.5, popadjust = TRUE, popupdate=
   epp.subp.input <- fnCreateEPPSubpops(epp.input, epp.subp, eppd)
 
   ## spectrum
-  demp <- read_specdp_demog_param(pjnz)
-  projp <- read_hivproj_param(pjnz)
+  demp <- read_specdp_demog_param(pjnz, use_ep5=use_ep5)
+  projp <- read_hivproj_param(pjnz, use_ep5=use_ep5)
   epp_t0 <- read_epp_t0(pjnz)
 
   ## If Urban/Rural fit, read percentage urban from EPP XML file
@@ -135,14 +135,14 @@ create_subpop_specfp <- function(projp, demp, eppd, epp_t0=setNames(rep(1975, le
 
 
 ## Prepare national fit. Aggregates ANC data from regional EPP files.
-prepare_national_fit <- function(pjnz, upd.path=NULL, proj.end=2013.5, hiv_steps_per_year = 10L){
+prepare_national_fit <- function(pjnz, upd.path=NULL, proj.end=2013.5, hiv_steps_per_year = 10L, use_ep5=use_ep5){
 
   ## spectrum
   if(!is.null(upd.path))
     demp <- read_demog_param(upd.path)
   else
-    demp <- read_specdp_demog_param(pjnz)
-  projp <- read_hivproj_param(pjnz)
+    demp <- read_specdp_demog_param(pjnz, use_ep5=use_ep5)
+  projp <- read_hivproj_param(pjnz, use_ep5=use_ep5)
   epp_t0 <- read_epp_t0(pjnz)
 
   specfp <- create_spectrum_fixpar(projp, demp, proj_end = as.integer(proj.end), time_epi_start = epp_t0[1], hiv_steps_per_year= hiv_steps_per_year)  # Set time_epi_start to match first EPP population