+++ NEWS.md
@@ -1,3 +1,7 @@
 # eppspectrum 0.1
 
 - Update to interface with `epp v0.3` which implements ANC routine testing (ANC-RT) likelihood in EPP 2017.
+
+# eppspectrum 0.1.1
+
+- Add option to use .ep5 file inputs instead of .DP file when preparing fit from PJNZ file. (use_ep5=FALSE by default.)